oncosense/requirements.txt
# OncoSense - MRI Brain Tumor Classification System
# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pytorch-lightning>=2.0.0

# Data Processing
kagglehub
pandas>=2.0.0
pyarrow>=12.0.0
imagehash>=4.3.0
Pillow>=9.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Image Augmentation
albumentations>=1.3.0
kornia>=0.7.0

# Calibration & Metrics
scipy>=1.10.0

# Explainability
pytorch-grad-cam>=1.4.0
captum>=0.6.0

# RAG Components
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
openai>=1.0.0

# UI & Reporting
streamlit>=1.28.0
reportlab>=4.0.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
